Create smaller kernels file
+Adaprt import in stencil.py
