+++ psydac/linalg/kernels/stencil2coo_kernels.py
@@ -1,6 +1,5 @@
 # coding: utf-8
 from pyccel.decorators import template
-from numpy import conjugate
 #========================================================================================================
 
 #__all__ = ['stencil2coo_1d_C','stencil2coo_1d_F','stencil2coo_2d_C','stencil2coo_2d_F', 'stencil2coo_3d_C', 'stencil2coo_3d_F']
@@ -191,173 +190,5 @@ def stencil2coo_3d_F(A:'T', data:'data', rows:'int64[:]', cols:'int64[:]',
     return nnz
 
 
-#========================================================================================================
-
-#!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!#
-#!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!#
-#!!!!!!! Conjugate on the first argument !!!!!!!#
-#!!!!!!!!!! This will need an update !!!!!!!!!!!#
-#!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!#
-@template(name='T', types=['float[:]', 'complex[:]'])
-def dot_product_1d(v1: 'T', v2: 'T', pads0: 'int64', shift0: 'int64'):
-    """
-    kernel for computing the inner product (case two complex 1d vectors)
-
-    Parameters
-    ----------
-        v1, v2 : 1d complex array
-            Data of the vectors from which we are computing the inner product.
-
-        pads0, shift0, shape0 : int
-            pads, shift and shape of the space of the vectors in the 0 direction.
-
-    Returns
-    -------
-        res : complex containing the results
-        """
-    #$omp parallel firstprivate( pads0, shift0) shared(res, v1, v2) private(shape0)
-    res = v1[0]-v1[0]
-    shape0, = v1.shape
-    #$omp for collapse(1) reduction(+ : res)
-    for i0 in range(pads0*shift0, shape0-pads0*shift0):
-        res += v1[i0].conjugate()*v2[i0]
-    #$omp end parallel
-    return res
-
-
-#========================================================================================================
-@template(name='T', types=['float[:,:]', 'complex[:,:]'])
-def dot_product_2d(v1: 'T', v2: 'T', pads0: 'int64', pads1: 'int64', shift0: 'int64',
-                              shift1: 'int64'):
-    """
-    kernel for computing the inner product (case two complex 2d vectors)
-
-    Parameters
-    ----------
-        v1, v2 : 2d complex array
-            Data of the vectors from which we are computing the inner product.
-
-        pads0, shift0, shape0, pads1, shift1, shape1 : int
-            pads, shift and shape of the space of the vectors in the 0/1 direction.
-
-    Returns
-    -------
-        res : complex containing the results
-    """
-    #$omp parallel firstprivate( pads0, pads1, shift0, shift1) shared(res, v1, v2) private(shape0, shape1)
-    res = v1[0, 0]-v1[0, 0]
-    shape0, shape1 = v1.shape
-    #$omp for collapse(2) reduction(+ : res)
-    for i0 in range(pads0*shift0, shape0-pads0*shift0):
-        for i1 in range(pads1*shift1, shape1-pads1*shift1):
-            res += v1[i0, i1].conjugate()*v2[i0, i1]
-    #$omp end parallel
-    return res
-
-
-#========================================================================================================
-@template(name='T', types=['float[:,:,:]', 'complex[:,:,:]'])
-def dot_product_3d(v1: 'T', v2: 'T', pads0: 'int64', pads1: 'int64', pads2: 'int64',
-                              shift0: 'int64', shift1: 'int64', shift2: 'int64'):
-    """
-    kernel for computing the inner product (case two complex 3d vectors)
-
-    Parameters
-    ----------
-        v1, v2 : 3d complex array
-            Data of the vectors from which we are computing the inner product.
-
-        pads0, shift0, shape0, pads1, shift1, shape1, pads2, shift2, shape2 : int
-            pads, shift and shape of the space of the vectors in the 0/1/2 direction.
-
-    Returns
-    -------
-        res : complex containing the results
-    """
-    #$omp parallel firstprivate( pads0, pads1, pads2, shift0, shift1, shift2) shared(res, v1, v2) private(shape0, shape1, shape2)
-    res = v1[0, 0, 0] - v1[0, 0, 0]
-    shape0, shape1, shape2 = v1.shape
-    #$omp for collapse(3) reduction(+ : res)
-    for i0 in range(pads0*shift0, shape0-pads0*shift0):
-        for i1 in range(pads1*shift1, shape1-pads1*shift1):
-            for i2 in range(pads2*shift2, shape2-pads2*shift2):
-                res += v1[i0, i1, i2].conjugate()*v2[i0, i1, i2]
-    #$omp end parallel
-    return res
-
-
-#========================================================================================================
-
-#Implementation of the axpy kernels
-
-@template(name='Tarray', types=['float[:]', 'complex[:]'])
-@template(name='T', types=['float', 'complex'])
-def axpy_1d(alpha: 'T', x: "Tarray", y: "Tarray"):
-    """
-        kernel for computing y=y+alpha*x
-
-        Parameters
-        ----------
-            x, y : 1d array
-                Data of the vectors from which we are computing the inner product.
-
-            alpha : scalar
-                Coefficient needed for the operation to multiply v2
-    """
-    n1, = x.shape
-    #$ omp parallel for default(private) firstprivate(alpha, n1) shared(x, y) schedule(static)
-    for i1 in range(n1):
-        y[i1] += alpha*x[i1]
-    return
-
-
-#========================================================================================================
-
-@template(name='Tarray', types=['float[:,:]', 'complex[:,:]'])
-@template(name='T', types=['float', 'complex'])
-def axpy_2d(alpha: 'T', x: "Tarray", y: "Tarray"):
-    """
-        kernel for computing y=y+alpha*x
-
-        Parameters
-        ----------
-            x, y : 2d complex array
-                Data of the vectors from which we are computing the inner product.
-
-            alpha : scalar
-                Coefficient needed for the operation to multiply v2
-    """
-    n1, n2 = x.shape
-    #$ omp parallel for default(private) firstprivate(alpha, n1, n2) shared(x, y) collapse(2) schedule(static)
-    for i1 in range(n1):
-        for i2 in range(n2):
-            y[i1, i2] += alpha * x[i1, i2]
-    return
-
-#========================================================================================================
-@template(name='Tarray', types=['float[:,:,:]', 'complex[:,:,:]'])
-@template(name='T', types=['float', 'complex'])
-def axpy_3d(alpha: 'T', x: "Tarray", y: "Tarray"):
-    """
-        kernel for computing y=y+alpha*x
-
-        Parameters
-        ----------
-            x, y : 3d complex array
-                Data of the vectors from which we are computing the inner product.
-
-            alpha : scalar
-                Coefficient needed for the operation to multiply v2
-    """
-    n1, n2, n3 = x.shape
-    #$ omp parallel for default(private) firstprivate(alpha, n1, n2, n3) shared(x, y) collapse(3) schedule(static)
-    for i1 in range(n1):
-        for i2 in range(n2):
-            for i3 in range(n3):
-                y[i1, i2, i3] += alpha*x[i1, i2, i3]
-    return
-
-
-#========================================================================================================
